# runtime/openapi/testdata/flow.json
{
  "openapi": "3.1.0",
  "info": { "title": "FlowsAPI", "version": "1" },
  "paths": {
    "/flows/json": {
      "get": {
        "operationId": "listFlows",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "flows": {
                      "type": "array",
                      "items": {
                        "type": "object",
                        "properties": {
                          "name": { "type": "string" },
                          "schedule": { "type": ["string", "null"] }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "data": {
                      "type": ["object", "null"],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "properties": {
                            "error": { "type": "string" },
                            "field": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/flows/json/{flow}": {
      "post": {
        "operationId": "startFlow",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": { "type": "object", "additionalProperties": true }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Flow Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Run" }
              }
            }
          },
          "400": {
            "description": "Flow Response Errors",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "data": {
                      "type": ["object", "null"],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "properties": {
                            "error": { "type": "string" },
                            "field": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "get": {
        "operationId": "getFlow",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { "$ref": "#/components/schemas/Run" }
                }
              }
            }
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "data": {
                      "type": ["object", "null"],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "properties": {
                            "error": { "type": "string" },
                            "field": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "limit",
          "in": "query",
          "required": false,
          "description": "",
          "schema": { "type": "number" }
        },
        {
          "name": "before",
          "in": "query",
          "required": false,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "after",
          "in": "query",
          "required": false,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "flow",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        }
      ]
    },
...
    "/flows/json/{flow}/{runId}/back": {
      "post": {
        "operationId": "backFlowRun",
        "responses": {
          "200": {
            "description": "Flow Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Run" }
              }
            }
          },
          "400": {
            "description": "Flow Response Errors",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "data": {
                      "type": ["object", "null"],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "properties": {
                            "error": { "type": "string" },
                            "field": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "flow",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "runId",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        }
      ]
    },
    "/flows/json/{flow}/{runId}/cancel": {
      "post": {
        "operationId": "cancelFlowRun",
        "responses": {
          "200": {
            "description": "Flow Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/Run" }
              }
            }
          },
          "400": {
            "description": "Flow Response Errors",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "data": {
                      "type": ["object", "null"],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "properties": {
                            "error": { "type": "string" },
                            "field": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "flow",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "runId",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        }
      ]
    },
...
    "/flows/json/{flow}/{runId}/{stepId}/callback": {
      "post": {
        "operationId": "callback",
        "requestBody": {
          "description": "",
          "content": {
            "application/json": {
              "schema": {
                "type": ["object", "string", "number", "boolean", "array"],
                "additionalProperties": true
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Callback Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": ["object", "string", "number", "boolean", "array"],
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Callback Response Errors",
            "content": {
              "application/json": {
                "schema": {
                  "properties": {
                    "code": { "type": "string" },
                    "data": {
                      "type": ["object", "null"],
                      "properties": {
                        "errors": {
                          "type": "array",
                          "properties": {
                            "error": { "type": "string" },
                            "field": { "type": "string" }
                          }
                        }
                      }
                    },
                    "message": { "type": "string" }
                  }
                }
              }
            }
          }
        }
      },
      "parameters": [
        {
          "name": "flow",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "runId",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "stepId",
          "in": "path",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "callback",
          "in": "query",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        },
        {
          "name": "element",
          "in": "query",
          "required": true,
          "description": "",
          "schema": { "type": "string" }
        }
      ]
    }
  },
  "components": {
    "schemas": {
      "Run": {
        "type": "object",
        "properties": {
          "config": {
            "type": "object",
            "properties": {
              "description": { "type": "string" },
              "stages": {
                "type": "array",
                "items": {
                  "type": "object",
                  "properties": {
                    "description": { "type": "string" },
                    "initiallyHidden": { "type": "boolean" },
                    "key": { "type": "string" },
                    "name": { "type": "string" }
                  },
                  "additionalProperties": false,
                  "required": ["key", "name"]
                }
              },
              "title": { "type": "string" }
            },
            "additionalProperties": false,
            "required": ["title"]
          },
          "createdAt": { "type": "string", "format": "date-time" },
          "data": { "type": ["object", "null"] },
          "id": { "type": "string" },
          "input": { "type": ["object", "null"], "additionalProperties": true },
          "name": { "type": "string" },
          "startTime": { "type": "string", "format": "date-time" },
          "startedBy": { "type": ["string", "null"] },
          "status": {
            "type": "string",
            "enum": [
              "NEW",
              "RUNNING",
              "AWAITING_INPUT",
              "FAILED",
              "COMPLETED",
              "CANCELLED"
            ]
          },
          "steps": {
            "type": "array",
            "items": { "$ref": "#/components/schemas/Step" }
          },
          "traceId": { "type": "string" },
          "updatedAt": { "type": "string", "format": "date-time" }
        },
        "required": [
          "id",
          "status",
          "name",
          "traceId",
          "createdAt",
          "updatedAt",
          "steps",
          "config",
          "input"
        ]
      },
...
      "Step": {
        "type": "object",
        "properties": {
          "createdAt": { "type": "string", "format": "date-time" },
          "endTime": { "type": ["string", "null"], "format": "date-time" },
          "error": { "type": ["string", "null"] },
          "id": { "type": "string" },
          "name": { "type": "string" },
          "runId": { "type": "string" },
          "stage": { "type": ["string", "null"] },
          "startTime": { "type": ["string", "null"], "format": "date-time" },
          "status": {
            "type": "string",
            "enum": ["PENDING", "FAILED", "COMPLETED", "CANCELLED"]
          },
          "type": { "type": "string", "enum": ["FUNCTION", "UI", "COMPLETE"] },
          "ui": { "$ref": "#/components/schemas/UiConfig" },
          "updatedAt": { "type": "string", "format": "date-time" },
          "value": {
            "type": [
              "string",
              "object",
              "array",
              "integer",
              "number",
              "boolean",
              "null"
            ],
            "additionalProperties": true
          }
        },
        "required": [
          "id",
          "runId",
          "status",
          "name",
          "type",
          "createdAt",
          "updatedAt",
          "value",
          "ui",
          "startTime",
          "endTime",
          "error"
        ]
      },
      "UiConfig": {
        "anyOf": [
          {
            "$ref": "#/components/schemas/__UiConfigSchemas/$defs/UiCompleteApiResponse"
          },
          {
            "$ref": "#/components/schemas/__UiConfigSchemas/$defs/UiPageApiResponse"
          }
        ]
      },
...
    }
  }
}